Subtask1 ending commit
Final commit to put up to date the GitHub repository. Several examples and tests have been taken to validate the correct funcionalities of our program with the features:
Export maze as a JSON, export maze as a JPEG image, import a JSON file representing a maze and exporting its corresponding maze.
A custom exception to check the consistency of the JSON file has been designed, as well as the several constant sued in the program have been taken into an independet file of constants.