Added goal and succesor functions, problem to JSON
There were added functions to define the goal and successors function for the problem in their corresponding files. In addition, a function to test the correct generation of random nodes to be inserted in the fronteir has been added. In export.py, a function to represent a problem as a JSON has been added. Finally, some minor changes to import statements in import_json have been made